+++ lanzou/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['api']
+__all__ = ['api', 'gui']
