+++ lanzou/gui/dialogs/__init__.py
@@ -1,0 +1,16 @@
+from lanzou.gui.dialogs.about import AboutDialog
+from lanzou.gui.dialogs.captcha import CaptchaDialog
+from lanzou.gui.dialogs.delete import DeleteDialog
+from lanzou.gui.dialogs.info import InfoDialog
+from lanzou.gui.dialogs.login import LoginDialog
+from lanzou.gui.dialogs.move import MoveFileDialog
+from lanzou.gui.dialogs.rec_folder import RecFolderDialog
+from lanzou.gui.dialogs.rename import RenameDialog
+from lanzou.gui.dialogs.setpwd import SetPwdDialog
+from lanzou.gui.dialogs.setting import SettingDialog
+from lanzou.gui.dialogs.upload import UploadDialog
+
+
+__ALL__ = ['LoginDialog', 'UploadDialog', 'InfoDialog', 'RenameDialog',
+           'SettingDialog', 'RecFolderDialog', 'SetPwdDialog', 'MoveFileDialog',
+           'DeleteDialog', 'KEY', 'AboutDialog', 'CaptchaDialog']
